Move Distribution commands from Artifactory
